Added theme file and topbar component
material-ui needs a theme to be setup in order to utilize many of the features, here I set a basic dark and light theme and added logic to hold the dark and light theme designation in a context to be utilized throughout the app. Also adding logic for the dark and light theme color palette switching upon calling the colorModes toggle color property.